Add Gate::authorize() to all Methods in Panel/ChapterController.php

+++ app/Http/Controllers/Panel/ChapterController.php
@@ -14,7 +14,7 @@ class ChapterController extends Controller
 {
     public function index()
     {
-//        Gate::authorize('view-episodes');
+        Gate::authorize('view-chapters');
 
         $chapters = Chapter::orderByDesc('id')->paginate(5);
         return view('panel.chapters.index',compact('chapters'));
@@ -22,7 +22,7 @@ class ChapterController extends Controller
 
     public function create()
     {
-//        Gate::authorize('create-episode');
+        Gate::authorize('create-chapter');
 
         $courses = Course::all();
         return view('panel.chapters.create',compact('courses'));
@@ -30,7 +30,7 @@ class ChapterController extends Controller
 
     public function store(CreateChapterRequest $request)
     {
-//        Gate::authorize('store-episode');
+        Gate::authorize('store-chapter');
 
         Chapter::create(
             $request->validated()
@@ -41,19 +41,23 @@ class ChapterController extends Controller
 
     public function showChapterEpisodes(Chapter $chapter)
     {
+        Gate::authorize('view-chapter-episodes');
+
         $episodes = $chapter->episodes()->paginate(5);
         return view('panel.chapters.chapter_episodes',compact('episodes'));
     }
 
     public function showChapterCourse(Chapter $chapter)
     {
+        Gate::authorize('view-chapter-course');
+
         $course = $chapter->course()->first();
         return view('panel.chapters.chapter_course',compact('course'));
     }
 
     public function edit(Chapter $chapter)
     {
-//        Gate::authorize('edit-episodes');
+        Gate::authorize('edit-chapters');
 
         $courses = Course::all();
         return view('panel.chapters.edit',compact(['chapter','courses']));
@@ -61,7 +65,7 @@ class ChapterController extends Controller
 
     public function update(UpdateChapterRequest $request, Chapter $chapter)
     {
-//        Gate::authorize('update-episodes');
+        Gate::authorize('update-chapters');
 
         $chapter->update(
             $request->validated()
@@ -72,7 +76,7 @@ class ChapterController extends Controller
 
     public function destroy(Chapter $chapter, Request $request)
     {
-//        Gate::authorize('delete-episodes');
+        Gate::authorize('delete-chapters');
 
         $chapter->delete();
         $request->session()->flash('status','فصل مورد نظر با موفقیت حذف شد !');
